Fixed code in docstring

+++ pyro/distributions/flipflow.py
@@ -23,7 +23,7 @@ class FlipFlow(Transform):
     >>> from pyro.distributions import InverseAutoregressiveFlow, FlipFlow
     >>> base_dist = dist.Normal(torch.zeros(10), torch.ones(10))
     >>> iaf1 = InverseAutoregressiveFlow(AutoRegressiveNN(10, [40]))
-    >>> ff = FlipFlow(torch.randperm(10))
+    >>> ff = FlipFlow(torch.randperm(10, dtype=torch.long))
     >>> iaf2 = InverseAutoregressiveFlow(AutoRegressiveNN(10, [40]))
     >>> iaf_dist = dist.TransformedDistribution(base_dist, [iaf1, ff, iaf2])
     >>> iaf_dist.sample()  # doctest: +SKIP
